# products/amliq/internal/shared/sdk/openapi-merged.yaml
openapi: 3.0.3
info:
  title: FinSavvy AI Platform API
  version: 1.0.0
  description: |
    Unified API for the FinSavvy AI enterprise fintech platform.
    Combines fraud detection, analytics, billing, ML retraining,
    onboarding, compliance, webhook, and audit services.
  contact:
    name: FinSavvy AI Engineering
    email: engineering@finsavvy.ai
    url: https://docs.finsavvy.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.finsavvy.ai/api/v1
    description: Production
  - url: https://sandbox.finsavvy.ai/api/v1
    description: Sandbox
security:
  - bearerAuth: []
...
paths:
...
  /onboarding/start:
    post:
      operationId: startOnboarding
      summary: Start a new tenant onboarding session
      tags: [Onboarding]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [company_name, email]
              properties:
                company_name:
                  type: string
                email:
                  type: string
                  format: email
                plan:
                  type: string
                  enum: [starter, professional, enterprise]
      responses:
        "201":
          description: Onboarding session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingSession"
...
  /onboarding/{id}/sandbox:
    post:
      operationId: provisionSandbox
      summary: Provision a sandbox environment for the tenant
      tags: [Onboarding]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "202":
          description: Sandbox provisioning started
          content:
            application/json:
              schema:
                type: object
                properties:
                  sandbox_url:
                    type: string
                    format: uri
                  api_key:
                    type: string
                  status:
                    type: string
...
  /analytics/transactions:
    get:
      operationId: getTransactionAnalytics
      summary: Get transaction analytics for a time range
      tags: [Analytics]
      parameters:
        - name: start
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Transaction analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionAnalytics"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /billing/invoices:
    get:
      operationId: listInvoices
      summary: List invoices for the current tenant
      tags: [Billing]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoices:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invoice"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (15 min expiry)
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Tenant-scoped API key
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
        request_id:
          type: string
...
    OnboardingSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
        company_name:
          type: string
        email:
          type: string
        plan:
          type: string
        current_step:
          type: integer
        total_steps:
          type: integer
        status:
          type: string
          enum: [in_progress, completed, abandoned]
        created_at:
          type: string
          format: date-time
...
    TransactionAnalytics:
      type: object
      properties:
        total_transactions:
          type: integer
        total_volume:
          type: number
        avg_amount:
          type: number
        fraud_rate:
          type: number
        approval_rate:
          type: number
        period_start:
          type: string
          format: date-time
        period_end:
          type: string
          format: date-time
...
    Invoice:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [draft, open, paid, void]
        due_date:
          type: string
          format: date
        pdf_url:
          type: string
          format: uri
...